% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topconfects.R
\name{confects_plot}
\alias{confects_plot}
\title{Top confident effect sizes plot}
\usage{
confects_plot(confects, n = 50)
}
\arguments{
\item{confects}{A "Topconfects" class object, as returned from limma_confects, edger_confects, etc.}

\item{n}{Number if items to show.}
}
\value{
A ggplot2 object. Working non-interactively, you must print() this for it to be displayed.
}
\description{
Top confident effect sizes plot
}

