% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_effects.R
\name{effect_contrast}
\alias{effect_contrast}
\title{Linear contrast effect.}
\usage{
effect_contrast(contrast)
}
\arguments{
\item{contrast}{Either a named vector with names matching the design matrix, or a vector of the same length as the number of columns in the design matrix. The effect size is the dot product of this with the fitted coefficients.}
}
\value{
An object defining how to calculate an effect size.
}
\description{
This is included for completeness and testing. limma_confects() provides a faster version.
}
