% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_contrast_ratio}
\alias{effect_contrast_ratio}
\title{Ratio of contrasts effect.}
\usage{
effect_contrast_ratio(contrast1, contrast2)
}
\arguments{
\item{contrast1}{First contrast weights, a vector with the same length as the number of columns in the design matrix.}

\item{contrast2}{Second contrast weights, a vector with the same length as the number of columns in the design matrix.

\code{sum(contrast2*beta)} should always produce a positive value.}
}
\value{
An object defining how to calculate an effect size.
}
\description{
Ratio of two contrasts, i.e. \code{sum(contrast1*beta)/sum(contrast2*beta)}.
}
