% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_gamma}
\alias{effect_gamma}
\title{Goodman and Kruskall's gamma, Yule's Q}
\usage{
effect_gamma(coef1, coef2)
}
\arguments{
\item{coef1}{Column numbers in the design matrix for the first condition, in some meaningful order.}

\item{coef2}{Corresponding column numbers for the second condition.}
}
\value{
An object defining how to calculate an effect size.
}
\description{
Goodman and Kruskall's gamma as an effect size. Yule's Q is a special case where coef1 and coef2 both have two coefficients, and is a symmetric effect size for the interaction of two experimental factors.
}
\details{
\code{effect_gamma_log2} is adapted to work with log2 scaled coefficients. This is almost certainly the version you want.
}
