% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_link_log2}
\alias{effect_link_log2}
\title{Adapt an effect size object to work with coefficients estimated from log2 transformed data or with a log2 link function}
\usage{
effect_link_log2(effect)
}
\arguments{
\item{effect}{An object definining an effect size on coefficients with untransformed scale.}
}
\value{
An object defining how to calculate an effect size from coefficients on a logarithmic scale.
}
\description{
Differential expression analysis usually fits coefficients to log2 transformed values, or with a log2 link function. This function adapts an effect size object to work with such log2-scale coefficients.
}
