% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_sd}
\alias{effect_sd}
\title{Standard-deviation of a set of coefficients as effect size}
\usage{
effect_sd(coef)
}
\arguments{
\item{coef}{The column numbers of the design matrix for the relevant coefficients.}
}
\value{
An object defining how to calculate an effect size.
}
\description{
This is intended as the effect size version of an ANOVA. The effect size is the standard deviation of some coefficients about their mean.
}

