% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_shift}
\alias{effect_shift}
\alias{effect_shift_log2}
\title{Shift of mass effect}
\usage{
effect_shift(coef1, coef2)

effect_shift_log2(coef1, coef2)
}
\arguments{
\item{coef1}{Column numbers in the design matrix for the first condition, in some meaningful order.}

\item{coef2}{Corresponding column numbers for the second condition.

effect_shift_log2 is adapted to work with log2 scaled coefficients. This is almost certainly the version you want.}
}
\value{
An object defining how to calculate an effect size.
}
\description{
Detect a "shift of mass" between two conditions. For example expression might move later in a time series in an experimental condition vs a control. If all expression shifted to a later time in the experimental condition, this would be given an effect size of 1. Conversely if all expression shifted to an earlier time, the effect size would be -1.
}

