% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_confects.R
\name{group_effect_rssm}
\alias{group_effect_rssm}
\title{Group effect for differential splicing}
\usage{
group_effect_rssm(design, coef, design_common = NULL)
}
\arguments{
\item{design}{Design matrix.}

\item{coef}{Column number in design matrix of the coefficient to be tested.}

\item{design_common}{Experimental! Optional sample-level design matrix. For example, this can be used to account for a batch effect or matched samples.}
}
\value{
A group effect object.
}
\description{
Create a group effect object to detect differential splicing (or similar). The effect size is the root sum of squared differences of the coefficient from the mean.
}
\details{
The coefficient should represent a difference between two conditions.

The idea is to detect differences in differential expression between features in a group (ie exons in a gene).
}
\seealso{
\code{\link{effect_rssm}}
}
