% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_confects.R
\name{group_effect_shift}
\alias{group_effect_shift}
\alias{group_effect_shift_stepup}
\alias{group_effect_shift_stepdown}
\alias{group_effect_shift_log2}
\alias{group_effect_shift_stepup_log2}
\alias{group_effect_shift_stepdown_log2}
\title{Group effect for shifts in 5' or 3' end usage}
\usage{
group_effect_shift(design, coef1, coef2, design_common = NULL)

group_effect_shift_stepup(design, coef1, coef2, design_common = NULL)

group_effect_shift_stepdown(design, coef1, coef2, design_common = NULL)

group_effect_shift_log2(design, coef1, coef2, design_common = NULL)

group_effect_shift_stepup_log2(design, coef1, coef2, design_common = NULL)

group_effect_shift_stepdown_log2(design, coef1, coef2, design_common = NULL)
}
\arguments{
\item{design}{Design matrix.}

\item{coef1}{Column number of coefficient for first condition in design matrix.}

\item{coef2}{Column number of coefficient for second condition in design matrix.}

\item{design_common}{Experimental! Optional sample-level design matrix. For example, this can be used to account for a batch effect or matched samples.}
}
\value{
A group effect object.
}
\description{
Create a group effect object to detect a shift in usage of features which are ordered within a group.
}
\details{
group_effect_shift_log2 is almost certainly the version you want.

The coefficients should represent the expression levels in two different conditions.

The _stepup and _stepdown versions may be used to look for shifts in start or end of transcription from RNA-seq or microarray data, where the observed levels are expected to be cumulative or reverse cumulative.
}
\seealso{
\code{\link{effect_shift}}, \code{\link{effect_shift_log2}}
}
