% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_confects}
\alias{limma_confects}
\title{Confident log2 fold changes using limma's treat function, linear effects only}
\usage{
limma_confects(fit, coef = NULL, fdr = 0.05, step = 0.01)
}
\arguments{
\item{fit}{A limma MArrayLM object.}

\item{coef}{Column number of coefficient or contrast to test.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of log2 fold changes to test.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
For all possible absolute log2 fold changes, which genes have at least this fold change at a specified False Discovery Rate?
}
\details{
\code{fit} should be produced using \code{lmFit}, optionally \code{contrasts.fit} if a contrast is needed, and then \code{eBayes}, as in normal limma usage.
}
