% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_limma.R
\name{limma_group_confects}
\alias{limma_group_confects}
\title{Group effect sizes starting from limma analysis}
\usage{
limma_group_confects(object, design, grouping, effect, intercor = 0,
  fdr = 0.05, step = 0.001)
}
\arguments{
\item{object}{An expression matrix or EList object, or anything limma's lmFit will accept.}

\item{design}{Design matrix.}

\item{grouping}{A data frame with at least two columns, one named "group" and one named "name". The "name" column should correspond to row names in \code{object}. The "group" column should provide a grouping of the named rows. The order within groups may be important.}

\item{effect}{A non-linear group effect size definition, created with one of the \code{group_effect_...} functions.}

\item{intercor}{Correlation of the contributions of different features to the effect size. 0 implies no correlation, 1 implies perfect correlation. Larger values will produce more conservative confects. Effect size estimates are unaffected.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of effect sizes to test.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
Fit a linear model to each feature using limma, then calculate non-linear effect sizes on groups of features that are a function of the fitted coefficients.
}
\details{
For RNA-Seq data, use of \code{limma::voom} is necessary to properly weight expression levels.
}
