% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_limma.R
\name{limma_nonlinear_confects}
\alias{limma_nonlinear_confects}
\title{Confident non-linear effect sizes using Wald tests and limma}
\usage{
limma_nonlinear_confects(object, design, effect, fdr = 0.05, step = 0.001)
}
\arguments{
\item{object}{An expression matrix or EList object, or anything limma's lmFit will accept.}

\item{design}{Design matrix.}

\item{effect}{A non-linear effect size definition, created with one of the \code{effect_...} functions.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of effect sizes to test.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
Fit a linear model to each gene using limma, then calculate non-linear effect sizes that are a function of the fitted coefficients.
}
\details{
For RNA-Seq data, use of \code{limma::voom} is necessary to properly weight expression levels.
}
