% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topconfects.R
\name{confects_plot_me}
\alias{confects_plot_me}
\title{Mean-expression vs effect size plot}
\usage{
confects_plot_me(confects)
}
\arguments{
\item{confects}{A "Topconfects" class object, as returned from limma_confects, edger_confects, etc.}
}
\value{
A ggplot2 object. Working non-interactively, you must print() this for it to be displayed.
}
\description{
Like plotMD in limma, but shows "confect" on the y axis rather than "effect" ("effect" is shown underneath in grey). This may be useful for assessing whether effects are only being detected only in highly expressed genes.
}

