% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger.R
\name{edger_confects}
\alias{edger_confects}
\title{Confident effect sizes using edgeR, both linear and non-linear}
\usage{
edger_confects(fit, coef = NULL, contrast = NULL, effect = NULL,
  fdr = 0.05, step = 0.01, null = "worst.case")
}
\arguments{
\item{fit}{An edgeR DGEGLM object.}

\item{coef}{Coefficient to test, as per \code{glmTreat}. Use either coef or contrast or effect.}

\item{contrast}{Contrast to test, as per \code{glmTreat}. Use either coef or contrast or effect.}

\item{effect}{A non-linear effect, created with one of the effect_... functions. Use either coef or contrast or effect.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of log2 fold changes to test.}

\item{null}{"null" parameter passed through to edger::glmTreat (if coef or contrast given). Choices are "worst.case" or "interval". Note that the default here is "worst.case", to be consistent with other functions in topconfects. This differs from the default for glmTreat.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.

Technical note when using a non-linear effect size: Signed confects are based on TREAT-style p-values. Unsigned consfects (generally with df>1) are based on comparing the best fit within the H0 region to the best fit overall, which may up to double p-values.
}
\description{
For all possible absolute log2 fold changes, which genes have at least this fold change at a specified False Discovery Rate?
}

