% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger.R
\name{edger_confects}
\alias{edger_confects}
\title{Confident log2 fold change based on the edgeR Quasi-Likelihood method}
\usage{
edger_confects(fit, coef = NULL, contrast = NULL, fdr = 0.05,
  step = 0.01, null = "worst.case")
}
\arguments{
\item{fit}{An edgeR DGEGLM object produced using \code{glmQLFit}.}

\item{coef}{Coefficient to test, as per \code{glmTreat}. Use either coef or contrast or effect.}

\item{contrast}{Contrast to test, as per \code{glmTreat}. Use either coef or contrast or effect.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of log2 fold changes to test.}

\item{null}{"null" parameter passed through to edger::glmTreat (if coef or contrast given). Choices are "worst.case" or "interval". Note that the default here is "worst.case", to be consistent with other functions in topconfects. This differs from the default for glmTreat.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
For all possible absolute log2 fold changes, which genes have at least this fold change at a specified False Discovery Rate?
}
