% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_confects.R
\name{edger_group_confects}
\alias{edger_group_confects}
\title{Group confects (differential exon usage, etc)}
\usage{
edger_group_confects(fit, group_id, group_effect, fdr = 0.05, step = 0.01,
  design = NULL)
}
\arguments{
\item{fit}{An edgeR DGEGLM object.}

\item{group_id}{A factor of length \code{nrow(fit)}, assigning items to groups (eg genes).}

\item{group_effect}{A \code{function(n)} returning an effect size object for a group with \code{n} members.}

\item{fdr}{False Discovery Rate to maintain.}

\item{step}{Step size when calculating confident effect sizes.}

\item{design}{The design matrix is usually taken from \code{fit}, however it may be overridden with this parameter. Note that dispersion estimates and df.prior are taken from \code{fit}. So you may estimate dispersions with a simpler model than is specified here, in order to at least obtain pessimistic confects when you lack replicates.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
Find differential exon usage, etc.
}
\details{
If the order of the members of a group is important, ensure the order of rows is correct in the original matrix.

Groups with less than two members will be ignored.

To construct a group design matrix, the design from \code{fit} will be repeated in a block diagonal matrix.
}
