% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_confects.R
\name{edger_group_confects}
\alias{edger_group_confects}
\alias{limma_group_confects}
\title{Group confects (differential 5' or 3' end usage, etc)}
\usage{
edger_group_confects(fit, group_id, group_effect, fdr = 0.05, step = 0.01)

limma_group_confects(object, group_id, group_effect, fdr = 0.05,
  step = 0.01, trend = FALSE)
}
\arguments{
\item{fit}{For edgeR, an edgeR DGEGLM object.}

\item{group_id}{A factor of length \code{nrow(fit)}, assigning items to groups (eg genes).}

\item{group_effect}{A group effect object created by one of the A \code{group_effect_...} functions.}

\item{fdr}{False Discovery Rate to maintain.}

\item{step}{Step size when calculating confident effect sizes.}

\item{object}{For limma, an expression matrix or EList object, or anything limma's lmFit will accept.}

\item{trend}{For limma, should \code{eBayes(trend=TRUE)} be used?}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
Find differential exon usage, etc.
}
\details{
If the order of the members of a group is important, ensure the order of rows is correct in the original matrix.

Groups with less than two members will be ignored.

To construct a group design matrix, the design from \code{fit} will be repeated in a block diagonal matrix.
}
