% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_contrast}
\alias{effect_contrast}
\title{Simple linear contrast effect.}
\usage{
effect_contrast(contrast)
}
\arguments{
\item{contrast}{A vector of the same length as the number of coefficients. The effect size is the dot product of this with the coefficients.}
}
\value{
An object defining how to calculate an effect size.
}
\description{
This is included for completeness and testing. limma_confects() and edger_confects() provide a faster version of this using limma's treat and edgeR's topTreat functions.
}
