% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effect_sd}
\alias{effect_rss}
\alias{effect_sd}
\title{Standard-deviation of a set of coefficients as effect size}
\usage{
effect_sd(coef)

effect_rss(coef)
}
\arguments{
\item{coef}{The column numbers of the design matrix for the relevant coefficients.}
}
\value{
An object defining how to calculate an effect size.

\code{effect_rss} may be better suited to comparing effect sizes from designs with differing numbers of coefficients, such as differential exon usage.
}
\description{
This is intended as the effect size version of an ANOVA. For effect_sd, the effect size is the standard deviation of some coefficients about their mean. For effect_rss, it is the root sum of squared differences from the mean.
}

