% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_confects}
\alias{limma_confects}
\title{Effect size magnitude for limma TREAT}
\usage{
limma_confects(fit, fdr = 0.05, max = 30, step = 0.05)
}
\arguments{
\item{fit}{A limma MArrayLM object with one coefficient. (Use limma::contrasts.fit reduce the fit object down to the contrast or coefficient that you are interested in.)}
}
\description{
Effect size magnitude for limma TREAT
}

