% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limma_confects}
\alias{limma_confects}
\title{Confident log2 fold changes using limma's treat function, linear effects only}
\usage{
limma_confects(fit, coef = NULL, contrast = NULL, fdr = 0.05, max = 30,
  step = 0.01)
}
\arguments{
\item{fit}{A limma MArrayLM object.}

\item{coef}{Coefficient to test, as per \code{glmTreat}. Use either coef or contrast, or have fit be the result of contrasts.fit.}

\item{contrast}{Contrast to test, as per \code{glmTreat}. Use either coef or contrast, or have fit be the result of contrasts.fit.}

\item{fdr}{False Discovery Rate to control for.}

\item{max}{Maximum log2 fold change to test for.}

\item{step}{Granularity of log2 fold changes to test.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
For all possible absolute log2 fold changes, which genes have at least this fold change at a specified False Discovery Rate?
}

