% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_confects.R
\name{limma_nonlinear_confects}
\alias{limma_nonlinear_confects}
\title{Confident non-linear effect sizes using limma}
\usage{
limma_nonlinear_confects(object, design, effect, fdr = 0.05, step = 0.01)
}
\arguments{
\item{object}{An expression matrix or EList object, or anything limma's lmFit will accept.}

\item{design}{Design matrix.}

\item{effect}{A non-linear effect, created with one of the effect_... functions.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of log2 fold changes to test.}
}
\value{
Technical note: Signed confects are based on TREAT-style p-values. Unsigned confects (generally with df>1) are based on comparing the best fit within the H0 region to the best fit overall, which may up to double p-values.

See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
Confident non-linear effect sizes using limma
}
