% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_confects.R
\name{nest_confects}
\alias{nest_confects}
\title{Find sets of discoveries for a range of effect sizes, controlling the False Discovery Rate for each set.}
\usage{
nest_confects(n, pfunc, fdr = 0.01, max = 30, step = 0.05)
}
\arguments{
\item{n}{Number of items being tested.}

\item{pfunc}{A function(indices, effect_size) to calculate p-values. Indices is a subset of 1:n giving the p-values to be computed.}

\item{fdr}{False Discovery Rate to control for.}

\item{max}{Maximum effect size to test for.}

\item{step}{Granularity of effect sizes to test.}
}
\description{
Find sets of discoveries for a range of effect sizes, controlling the False Discovery Rate for each set.
}

