% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_confects}
\alias{normal_confects}
\title{Confident effect sizes from from normal or t distributions}
\usage{
normal_confects(mean, sd = 1, df = Inf, fdr = 0.05, step = 0.01)
}
\arguments{
\item{mean}{A vector of means.}

\item{sd}{A single or vector of standard deviations (or if t distribution, scales).}

\item{df}{A single or vector of degrees of freedom, for t-distribution. Inf for normal distribution.}

\item{fdr}{False Discovery Rate to control for.}

\item{step}{Granularity of effect sizes to test.}
}
\value{
See \code{\link{nest_confects}} for details of how to interpret the result.
}
\description{
A general purpose confident effect size function for where a normal or t distribution can be assumed. Calculates confident effect sizes based on a mean and standard deviation (normal distribution) or mean and scale (t distribution).
}
